resize handling, async error handling
with the introduction of props.render(), now, resize() does not fire
when page first loads. only render() will run once. and for
render-on-demand, use render prop to control manually.

also, it fixes render running when dimensions is set, which was bad.

==
async error was ambiguous even for a simple syntax error. now, it uses
console.error() for more detailed error printing.