add `props.exportFrames`
Works the same ways as pressing CMD+SHFT+s to export video/frames. It allows exporting with a custom key when hotkeys are disabled. It may also be useful in situations where the keyboard is not available (ie. mobile)

Note that calling the function again while recording will finish the
current recording.