fix `loopCount` jumping
it used to check for `prevFrame > frame` or `prevFrame >= frame`
depending on playing or recording, but `frame` is unreliable as it may
or may not increment by 1 or at all - to keep up with time. so, there
was a glitch sometimes.

now, it simply checks for `frame === 0` to advance the `loopCount`.