fix: use `hot.off()`
for vite communication for ssam plugins, there used to be a `hotReloaded` flag to check if it's a first hot reload. it was to make sure ws listeners are only created once as there were no way for Vite to turn off listeners. now Vite has `hot.off()`, so ssam now can refer to each listener callback and remove them properly at each hot reload. for this to work, git success callback needed to be wrapped as it relies on sketch settings and states, not just returned `data` from the plugins.