fix!: remove `settings.centered`
`centered` was doing canvas centering and fit to parent (window) at the same time. it was also adding flex centering css inline style to canvas parent. this can lead to issues in custom dom set up. it is not removed. i should instead add `scaleCanvas` or `fitToParent` settings for control.