fix!: handling fit to parent
when there is a custom parent, ssam would look at grand prent element and add scaling to parent element. this can lead to custom CSS issues. there should be a minimum amount of CSS code within Ssam, and there should be a way to turn them all off. It's better handled in CSS outside of Ssam.