fix: fullscreen canvas resizing
resize handler only ran if fullscreen and ssam's own canvas, but changed to any canvas that is fullscreen. I've noticed that the responsive fullscreen canvas didn't work when using with React as I had to use <canvas ref={}> to connect ssam to an existing canvas. I either had to manage resizing manually within the component (which requires more ssam code changes) or just let the current handler to run in any fullscreen canvas situation and decided go with the latter.