fix: remove Wrap.create
`Wrap.create()` static method was introduced in `0.19.0` to simplify
typing and casting class properties. But it had an unintentional effect
of resizing method not properly registered and caused canvas stretching
at fullscreen canvas resizing and a few other nasty bugs. (ex.
flickering, streching of canvas, webgpu resize error) I believe it is
due to object instance not properly instantiated if done within a class
static method.

`Wrap.create()` did not really solve or add anything except for making
easier to handle typescript. In this commit, I reverted back to how it
was in `0.18.5` in terms of handling the creation of the new `Wrap`
object.

Maybe, I will revisit this issue later if I find an extra value.